requirements.txt
beautifulsoup4==4.12.3
PyYAML==6.0.2
requests==2.32.3
pytest==8.3.5

